Move get_version_info() to __init__.py
