+++ fx_crash_sig/cmd_get_crash_sig.py
@@ -9,7 +9,9 @@ import sys
 
 import ujson as json
 
-from fx_crash_sig.crash_processor import CrashProcessor, get_version_info
+from fx_crash_sig import get_version_info
+from fx_crash_sig.crash_processor import CrashProcessor
+
 
 DESCRIPTION = """
 Takes raw crash trace and symbolicates it to return the crash signature
